refactor(infra): ajustar imports e referências após reestruturação de pacotes
